Move DefCollector to its own module.
